Refactor to use mounts instead of symlinks and resolve symlinkers in emulator (#216)
This is a big PR that achieves three things:

## Remove our special dir symlinking hacks
In the past, /dev, /proc, /sys, /run, /tmp would be symlinked inside the
rootfs. This would lead to problems. You could detect that those
directories are not mounts. Also you'd have to do some hacking so that
path resolution on those paths wouldn't recurse infinitely.

Mounting is clearly a better solution. The problem is that mounting
requires administrator permissions. This is why we now use a privileged
executable called felix86-mounter, which felix86 passes a rootfs to, and
it returns a temporary path in /run, that has mounted the rootfs itself
(/ needs to be a mount, so thats a nice bonus), and /dev & co inside.

## Resolve symlinks instead of relying on the kernel to do that
There was a bug in our symlink implementation that modified the source
path to have the rootfs. This *happened to work*, until you have to
worry about chroots. So now we don't modify the source path, and resolve
the symlink ourselves.

## Add chroot syscall and tests
Added a chroot syscall implementation that just changes
g_config.rootfs_path to the new root