Check if landing blocks overwrite our flags (#324)
Introduces a new option, `FELIX86_SCAN_AHEAD_MULTI` (disabled by
default, to be enabled in the future) which will scan the first few
instructions of a jump target, if we know where it lands, to see if it
overwrites the flags so as to omit calculating them. Should improve
performance in tight loops.